+++ src/main/java/com/definesys/dsgc/service/esbHome/bean/EsbHomeCard.java
@@ -8,12 +8,11 @@ package com.definesys.dsgc.service.esbHome.bean;
  * @Version 1.0
  **/
 public class EsbHomeCard {
-//apiTotal，sysTotal，userTotal，apiVisitTotal
 
 
     private Integer total;
-    private Integer weekRate;//百分比
-    private Integer dayRate;//百分比
+    private Double weekRate;//百分比
+    private Double dayRate;//百分比
     private Integer dataAdd;
 
 
@@ -25,19 +24,19 @@ public class EsbHomeCard {
         this.total = total;
     }
 
-    public Integer getWeekRate() {
+    public Double getWeekRate() {
         return weekRate;
     }
 
-    public void setWeekRate(Integer weekRate) {
+    public void setWeekRate(Double weekRate) {
         this.weekRate = weekRate;
     }
 
-    public Integer getDayRate() {
+    public Double getDayRate() {
         return dayRate;
     }
 
-    public void setDayRate(Integer dayRate) {
+    public void setDayRate(Double dayRate) {
         this.dayRate = dayRate;
     }
